add config files
